Add worker and shift related dataclasses to core NPT-58 (#32)
* Introduce core for worker related classes.

Creation of dataclasses to define types for worker, worker_dimension (formerly worker_param) and worker_property. These dataclasses are used to create objects used in the core. The databases now return object instances of these dataclasses instead of mongo documents.
Of note, these changes required a change in the FE to account for the fact that worker id are now "id" instead of "_id". This broke the edit functionality of the shifts as they use the same template. To be fixed in next commit.

* Build core for shift, shift_dimension and shift_property.

* Fix delete shift dimension route, was not deleting the associated shift properties correctly.

---------

Co-authored-by: Felipe Kharaba <felipekharaba@Felipes-MacBook-Pro.local>