Add schedule tab and connect to engine NPT 60 (#39)
* Change engine outputs content, add core schedule and add solver route.

This provides the minimum BE to send schedule to FE. At the moment, schedule is directly sent to FE and is not stored in database. This will be implemented in a future PR.

* Add Schedule tab in FE.

This is the first draft of the schedule tab. This is still wip, require additional work on column/row data structure and date display.

* Improve date format and fix checks errors.

---------

Co-authored-by: Felipe Kharaba <felipekharaba@Felipes-MacBook-Pro.local>