Rebuild google constraints NPT-88 (#55)
* Build draft basic custom constraint input.

Build the BE core and FE for sum and sequence constraints. Engine needs some modifications. BE and FE not connected yet.

* Add functionalities to read, update (active only) and delete constraints.

This includes the FE and the BE for sum constraint only. Code to be able to udpate full constraint still to be implemented. Engine still not connected.

* Change engine to new constraint object.

This includes changes to engine's input dataclasses, the model and the tests.

* Connect engine, and add sequence and order constraints.

* Fix check errors.

Some of the errors where fixed by either ignoring them (issue with str to Literal) or forced casting (need to add types on constraints build blocks).

* Fix constraint fetch for engine to active only.

* Run make format.

* Add all files.

I was in the wrong directory when adding files to the stagging area.

* Run make format.

* Comment out model save to txt.

* Fix name for requestStore.ts in git tracked files list.

---------

Co-authored-by: Felipe Kharaba <felipekharaba@Felipes-MacBook-Pro.local>