Add basic framework for contraint input (#131)
* Add first draft nlp for constraint input.

Constraint are entered by the user as text. Backend parse text with Spacy, and map it to core constraint format. Very much WIP.
First approach was to use NER only, with custom patterns to identify keywords. This proved to be limited, particularly when several shifts, or a shift and a worker are mentionned, as relation to other entities is unknown.
Second and current approach is still use NER, but this time analyse the hierarchy of the dependencies between them. This works well when only one shift, or a shift and a worker are mentionned. It is limited when two shifts are mentionned (e.g. shift A before shift B).
Third approach, to come, will therefore be analyse the dependencies based on their nature to analyse sense of the sentence (who, what, when, etc), combine with out NER to isolate information of interest.

Also include a first test data set with 10+ examples. Each example include the input sentence, the expected nlp parsing output, and the expected mapping to constraint output.

Mapping to constraint not built for now (there is a very rudimentary first draft, but not working).

* New NLP WIP.

* Start building constraint entry with templates.

Draft FE only for now.

* Add template constraint, and first draft of constraint input.

Next step is to connect the constraint object to the template usage to create a constraint.

* Add basic sum, seq and ord constraints. Still wip.

* Add mapping for sum, seq and ord constraints.

* Restructure constraint mapping.

* Finalize first iteration of basic constraint input.

This is still WIP, additional functionalities and templates to come in subsequent PRs.

* Fix check error.

* Fix check error.

---------

Co-authored-by: Felipe Kharaba <felipekharaba@Felipes-MacBook-Pro.local>