Change shift and workers options to a list of options instead of a dict.
Create a dataclass ShiftWorkerOption for the type of the shift and worker options. Add a field category_name to ShiftWorkerOption to contain the name of the category in order to avoid it being changed when camelizing the object. Options for shift and worker blocks in the constraint template are a list of ShiftWorkerOption.
Of note, stats is also using ShiftWorkerOption for the selection of shifts.