Change frontend constraint creation to use ShiftWorkerOption objects.
For worker and shift blocks, the template is a list of ShiftWorkerOptionT objects. For boolean dimensions, the ShiftWorkerOptionT object is duplicated, with one with name value true, and the other with name value false. The this new list is converted into a dictionary with keys the different categoryName contained in the ShiftWorkerOptionT list, and value an array of the ShiftWorkerOptionT with the same categoryName. This is used to build the select object. In the select object, names for boolean dimensions are converted to "categoryName" if name is true, "Not categoryName" if name is false. When selecting an option, the selected ShiftWorkerOptionT is added to the constraint's block field.