Rebuild constraint build (#413)
* Add key "is_bool_dim" to worker and shift options in constraint templates.

* Change shift and workers options to a list of options instead of a dict.

Create a dataclass ShiftWorkerOption for the type of the shift and worker options. Add a field category_name to ShiftWorkerOption to contain the name of the category in order to avoid it being changed when camelizing the object. Options for shift and worker blocks in the constraint template are a list of ShiftWorkerOption.
Of note, stats is also using ShiftWorkerOption for the selection of shifts.

* Rearrange constraint tab components folders in FE.

* Change frontend constraint creation to use ShiftWorkerOption objects.

For worker and shift blocks, the template is a list of ShiftWorkerOptionT objects. For boolean dimensions, the ShiftWorkerOptionT object is duplicated, with one with name value true, and the other with name value false. The this new list is converted into a dictionary with keys the different categoryName contained in the ShiftWorkerOptionT list, and value an array of the ShiftWorkerOptionT with the same categoryName. This is used to build the select object. In the select object, names for boolean dimensions are converted to "categoryName" if name is true, "Not categoryName" if name is false. When selecting an option, the selected ShiftWorkerOptionT is added to the constraint's block field.

* Rebuild constraint build.

When creating a constraint, a constraint build object is saved to the database. Then its text, missing properties and active are built, and send back to the frontend, they are not saved to the database as they are dependent on the values of workers and shifts that can change.

* Change get, update and delete constraint build to the new methods.

* Fix pylint and mypy check errors.

Tests are not working, need to fix constraint parser.

* Amend the constraint parser to use the new constraint build format.

Of note, I now need to align the shift selector for stats as it uses the constraint parser.

* Align stats code to ShiftWorkerOption in BE.

Test are still failing, this is due to the test data for the constraint parser that is not consistent.

* Fix constraint parser test data.

No check error in BE. Still need to check format of shift_dim_dict fed into constraint parser in BE, and make required changes to FE for stats to use ShiftWorkerOptionT.

* Change shift_dim_dict and worker_dim_dict key for boolean dimensions to the boolean value, instead of "the name of the dimension" and "not the name of the dimension" for constraint parser.

* Change stats in FE to use ShiftWorkerOptionT.

Checks not run yet.

* Fix FE check errors related to ShiftWorkerOptionT adoption.

There are still FE check errors, but they are due to the misshandling of the new format for missing properties.

* Add missing properties component to FE.

All FE check pass.

* Fix issue in BE where constraint with reference or relative shift missing were not deactivated.

* Align deactivated constraint format in campaign tab with the one in constraint tab.

All checks OK.

* Comment out method deleting worker or shift from constraint when deleting.

This needs to be handled in next PR.

* Change poetry version to 1.8.2 (same as local) in backend_check.yml

---------

Co-authored-by: Felipe Kharaba <felipekharaba@Felipes-MacBook-Pro.local>